+++ runtime/src/governance/mod.rs
@@ -1,0 +1,2 @@
+pub mod collective_management;
+pub mod tracks;
